Big Improvements
Abstracted the collision detection and made it more functional. It doesn't contain the complex rotation mechanics of Tetris but it avoids any errors. Gained the ability to clear lines. Added quality features like a small delay when turning left or right. Still left todo is adding a held piece and future pieces preview.